+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 1.4.0
-- Custom move generation and processing:
+- Custom move generation and processing (beta):
   - `MoveGenerator`: optionally supply a list of these in `Variant.moveGenerators`. These can be used to generate custom moves based on the state. You can generate both normal move types (e.g. add an extra move to a piece on a certain square or something), or totally different types of move that can be managed by `MoveProcessor`.
   - `MoveProcessor`: supply a list of these in `Variant.moveProcessors` to process custom move types.
 - `StateTransformer`: a way to transform game states based on player perspectives.